refactor: Move conversion methods and time formatting to shared code
- Move TeeTimeDocument/TeeTimeSlot conversion methods to TeeTimes feature module
- Add formattedTime extension property on LocalTime to core:extensions
- Add formattedTimes property on TeeTime model using shared extension
- Create reusable TtcTimePickerDialog in Android ui/common
- Update Android and iOS code to use shared time formatting
- Add core:extensions dependency to core:models

This addresses PR #4 feedback by consolidating duplicate formatTime
implementations and moving document conversion to the appropriate layer.
